Calculate size composition across aggregated subareas

Calculate size composition across aggregated subareas

calc_sizecomp_subareas(racebase_tables, size_comps)

Arguments

racebase_tables

data object created from `gapindex::get_data()``

size_comps

a dataframe of stratum biomass, result object from either gapindex::calc_sizecomp_aigoa_stratum() or gapindex::calc_sizecomp_bs_stratum()

Value

Field	Description

YEAR	Year the survey was conducted in.

AREA_ID	Area ID code for each statistical area used to produce production estimates (e.g., biomass, population, age comps, length comps). Each area ID is unique within each survey.

LENGTH_MM	Length of a specimen in millimeters.

POPULATION_COUNT	The estimated population caught in the survey for a species, group, or total for a given survey.

SEX	Sex of a specimen where "1" = "Male", "2" = "Female", "3" = Unsexed.

SPECIES_CODE	The species code of the organism associated with the 'common_name' and 'scientific_name' columns. For a complete species list, review the code books.

SURVEY_DEFINITION_ID	This number uniquely identifies a survey. Name and description of survey. The column 'survey_id' is associated with the 'srvy' and 'survey' columns. For a complete list of surveys, review the code books.