Calculate index of total biomass across aggregated subareas

Calculate index of total biomass across aggregated subareas

Usage

calc_biomass_subarea(racebase_tables = NULL, biomass_strata = NULL)

Arguments

racebase_tables

data object created from `gapindex::get_data()``

biomass_strata

a dataframe of stratum biomass, result object from gapindex::calc_biomass_stratum()

Value

Field	Description

YEAR	Year the survey was conducted in.

AREA_ID	Area ID code for each statistical area used to produce production estimates (e.g., biomass, population, age comps, length comps). Each area ID is unique within each survey.

BIOMASS_MT	The estimated biomass caught in the survey for a species, group, or total for a given survey.

BIOMASS_VAR	The estimated biomass variance caught in the survey for a species, group, or total for a given survey.

N_WEIGHT	Total number of hauls with positive catch/weighed taxon data used in calculation.

N_COUNT	Total number of hauls with positive taxon counts used in calculation.

N_HAUL	Total number of valid hauls used in calculation.

N_LENGTH	Total number of hauls with taxon length data used in calculation.

CPUE_KGKM2_MEAN	The mean of catch weight (kilograms) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.

CPUE_KGKM2_VAR	The variance of mean of catch weight (kilograms) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.

CPUE_NOKM2_MEAN	The mean of catch count (number) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.

CPUE_NOKM2_VAR	The variance of mMean of catch count (number) divided by area (squared kilometers) swept by the net used in design-based indicie calculation.

POPULATION_COUNT	The estimated population caught in the survey for a species, group, or total for a given survey.

POPULATION_VAR	The estimated population variance caught in the survey for a species, group, or total for a given survey.

SPECIES_CODE	The species code of the organism associated with the 'common_name' and 'scientific_name' columns. For a complete species list, review the code books.

SURVEY	Name and description of survey. The column 'survey' is associated with the 'srvy' and 'survey_id' columns.

SURVEY_DEFINITION_ID	This number uniquely identifies a survey. Name and description of survey. The column 'survey_id' is associated with the 'srvy' and 'survey' columns. For a complete list of surveys, review the code books.