Calculate numerical CPUE at length by haul

Uses Equation 16 in Wakabayashi et al. 1985 to calculate numbers-catch-per-unit-effort (CPUE, Numbers/km2) at length.

Usage

calc_sizecomp_stratum(
  racebase_tables = NULL,
  racebase_cpue = NULL,
  racebase_stratum_popn = NULL,
  spatial_level = c("stratum", "haul")[1],
  fill_NA_method = c("AIGOA", "BS")[1]
)

Arguments

racebase_tables

data object created from gapindex::get_data().

racebase_cpue

object created from gapindex::calc_cpue().

racebase_stratum_popn

object created from gapindex::calc_biomass_stratum()

spatial_level

string, one of c("stratum", "haul"). Should size compositions be calculated at the level of the "stratum" (used for standard size compositions) or "haul" (used for ModSquad model-based data-inputs).

fill_NA_method

string, one of c("AIGOA", "BS"). This argument changes the way hauls with positive weights but no associated size data are dealt with. In the EBS, NBS, and BSS survey regions ("BS"), these hauls contribute to the dummy length -9 category for their respective strata. In the AI and GOA survey regions ("AIGOA"), an average size distribution is applied to these hauls so the length -9 category does not exist in the AI or GOA versions of the size composition.

Value

dataframe of numerical CPUE ("S_ijklm_NO_KM2") by survey ("SURVEY"), year ("YEAR"), stratum ("STRATUM"), haul ("HAULJOIN), species (SPECIES_CODE), sex (SEX), and length (LENGTH_MM).